requirements.txt
vosk>=0.3.45
pyttsx3>=2.90
SpeechRecognition>=3.10.0
requests>=2.31.0
openai>=1.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
# Audio (Linux/Mac - pyaudio; Windows - optional)
pyaudio>=0.2.14; sys_platform != "win32"
# Image processing (for GIF animation)
Pillow>=10.0.0

